+++ browser.js
@@ -66,6 +66,8 @@ export {
   getPluginStatus,
   clearPlugins,
   validatePlugin,
+  collectPluginModuleCapabilities,
+  listPluginModuleCapabilities,
 } from './plugins/index.js';
 
 import {